requirements.txt
numpy==1.20.3
scikit_image==0.16.2
runstats==2.0.0
pytorch_lightning==1.4.7
h5py==2.10.0
PyYAML==5.4.1
torchmetrics==0.5.1
